+++ R/SMF.R
@@ -84,8 +84,8 @@
 #' @param opt_identifier `character` with the identifier to be used for optional
 #'     columns passed through `...`. This is used to create the column names for
 #'     optional columns in the format `opt_{opt_identifier}_{column_name}`. If
-#'     not provided, the default is `"global"`. Must the same length of the opt
-#'     columns provided in `...` or `1L`.
+#'     not provided, the default is `"global"`. Must be of length 1 or matching
+#'     the number of opt columns provided in `...`.
 #'
 #' @param ... Additional optional columns to add. These arguments must be named.
 #'   The function will automatically prepend `"opt_"` to the names if not