+++ NEWS.md
@@ -3,7 +3,8 @@
 ## Changes in version 0.97.16
 
 - Update vignette with high-level functionality to export a mzTab-M file.
-- Add helper functions `mzTabMToList()` and `.NAtonull()`.
+- Add helper functions `.NAtonull()`.
+- Add `as.list()`.
 - Fix bug in instrument field sorting.
 - Support `identifier` for optional columns in SML, SMF, SME.
 - Update `readMzTabM()` to create a `MzTabM` object.